[web] Add search
