+++ web/src/utils.js
@@ -162,6 +162,36 @@ const useLocalStorage = (storageKey, fallbackState) => {
   return [value, setValue];
 };
 
+/**
+ * Kudos to Sumit kumar Singh.
+ * See https://designtechworld.medium.com/create-a-custom-debounce-hook-in-react-114f3f245260
+ */
+
+const useDebounce = (callback, delay) => {
+  const timeoutRef = useRef(null);
+
+  useEffect(() => {
+    // Cleanup the previous timeout on re-render
+    return () => {
+      if (timeoutRef.current) {
+        clearTimeout(timeoutRef.current);
+      }
+    };
+  }, []);
+
+  const debouncedCallback = (...args) => {
+    if (timeoutRef.current) {
+      clearTimeout(timeoutRef.current);
+    }
+
+    timeoutRef.current = setTimeout(() => {
+      callback(...args);
+    }, delay);
+  };
+
+  return debouncedCallback;
+};
+
 const hex = (value) => {
   const sanitizedValue = value.replaceAll(".", "");
   return parseInt(sanitizedValue, 16);
@@ -248,6 +278,7 @@ export {
   classNames,
   useCancellablePromise,
   useLocalStorage,
+  useDebounce,
   hex,
   toValidationError,
   locationReload,
